Problem-types moved to pages
problem types are pages
