+++ src/App.js
@@ -16,9 +16,9 @@ import Problems from './pages/nav/problems'
 import SignUp from './pages/nav/signup'
 import ProblemDetail from './pages/problems/problemDetail'
 import AddProblems from './pages/nav/addProblems'
-import ICPC from './components/problems/problem-types/ICPC'
-import Interview from './components/problems/problem-types/Interview'
-import Programming from './components/problems/problem-types/Programming'
+import ICPC from './pages/problems/problem-types/ICPC'
+import Interview from './pages/problems/problem-types/Interview'
+import Programming from './pages/problems/problem-types/Programming'
 
 function App() {
   return (
