+++ src/pages/problems/problem-types/Interview.js
@@ -4,7 +4,7 @@
  */
 
 import React from 'react'
-import { InterviewFilter } from './problem-filters/InterviewFilter'
+import { InterviewFilter } from '../../../components/problems/problem-filters/InterviewFilter'
 
 function Interview() {
   return (
